+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## CHANGELOG (4/26/2021 11:42 am)
 
-1. removed Post components horizontal scrollbar.
+1. added thumbnail image support.
+2. added image support in focus mode (high res version).